+++ src/vue/components/containment/navigation-drawer/gk-navigation-drawer-types.ts
@@ -1,0 +1,14 @@
+export const gkNavigationDrawerLocations = [
+  'start',
+  'end',
+  'left',
+  'right',
+  'top',
+  'bottom',
+] as const
+
+export type GkNavigationDrawerLocation = (typeof gkNavigationDrawerLocations)[number]
+
+export type GkNavigationDrawerImageSlotProps = {
+  image: string | undefined
+}